Refactor combobox/comboboxmulti
